Fix `process.cwd()` path issue in Windows
Node.js's `path` methods return `\\` paths in Windows instead of the
usual `/` in Unix based OSs.

Because of that, when running `cmz` in Windows the generated class names
would actually contain the entire path of the current file instead of
just the relative path of the project as is expected.

This fixes this issue by transforming the returning path from
`process.cwd` into a Unix-like form.

```js
const root = process.cwd()
// Unix => foo/bar
// Windows => foo\\bar

slash(root)
// Unix => foo/bar
// Windows => foo/bar
```